Check that the file exists before deleting.

+++ Site/SiteVideoMediaLocalMover.php
@@ -93,7 +93,9 @@ class SiteVideoMediaLocalMover extends SiteVideoMediaMover
 
 	protected function cleanUp($path)
 	{
-		unlink($path);
+		if (file_exists($path)) {
+			unlink($path);
+		}
 	}
 
 	// }}}
